Fix table_name_prefix
In some circumstances the table_name_prefix is not loaded
correctly from rails. If this happens the users auth_tokens
can not be loaded.

I actually do not why this does not work, we do not even have
an ApplicationRecord in our app. Anyhow the current setup is a bit
overcomplicated and does not even take into account that an app
might have an ApplicationRecord.

I removed the top level ApplicationRecord as this does not make any
sense in this gem. The namespaced ApplicationRecord has been renamed
to BaseRecord, since ApplicationRecord is reserved for Apps and should
not be used in Engines IMO.